Create full moon
In this program, we use the ephem module to calculate the date of the next full moon. We also use the smtplib module to send email reminders when there is a full moon. We define a function called full_moon_reminder() that calculates the date of the next full moon and sends an email reminder. We then define a Flask route that displays the next full moon date on the web page. We also check if the next full moon is within the next 24 hours and call the full_moon_reminder() function if necessary. Finally, we start the Flask application using the run() method.